fastAPI/requirements.txt
fastapi==0.111.1
pydantic>=2.7.0
mysql-connector-python
typing